Moving the classes to new packages
