+++ src/test/java/test/testClasses/UsersTest.java
@@ -3,7 +3,7 @@ package test.testClasses;
 import test.BaseTest;
 import org.testng.annotations.Test;
 import org.testng.Assert;
-import test.clients.UsersClient;
+import test.clients.users.UsersClient;
 
 public class UsersTest extends BaseTest {
